Move check_connection to utils
