+++ sources/rest_api/utils.py
@@ -1,6 +1,9 @@
 from functools import reduce
 from operator import getitem
-from typing import Any, Sequence, Union
+from typing import Any, Sequence, Union, Tuple
+
+from dlt.common import logger
+from dlt.extract.source import DltSource
 
 
 def join_url(base_url: str, path: str) -> str:
@@ -13,3 +16,15 @@ def create_nested_accessor(path: Union[str, Sequence[str]]) -> Any:
     if isinstance(path, (list, tuple)):
         return lambda d: reduce(getitem, path, d)
     return lambda d: d.get(path)
+
+
+def check_connection(
+    source: DltSource,
+    *resource_names: str,
+) -> Tuple[bool, str]:
+    try:
+        list(source.with_resources(*resource_names).add_limit(1))
+        return (True, "")
+    except Exception as e:
+        logger.error(f"Error checking connection: {e}")
+        return (False, str(e))
